feat: OpenAPI 生成物を管理する
utoipa の生成結果を docs/openapi.json としてコミットし、CI で再生成差分を検出できるようにする。
