+++ README.md
@@ -14,9 +14,15 @@
 
 ## API定義
 
-Seichi Portal ではフロントエンドとバックエンド間の通信に REST API を使っており、API のスキーマは [OpenAPI v3.0.0](https://spec.openapis.org/oas/v3.0.0) ベースの[スキーマ管理用リポジトリ](https://github.com/GiganticMinecraft/seichi-portal-api-schema)に配置し、管理しています。
+Seichi Portal ではフロントエンドとバックエンド間の通信に REST API を使っており、API のスキーマは [OpenAPI v3.0.0](https://spec.openapis.org/oas/v3.0.0) ベースの [`docs/openapi.json`](./docs/openapi.json) を正本として管理しています。
 
-API 定義は GitHub Pages 上に[公開](https://giganticminecraft.github.io/seichi-portal-api-schema/)しており、すぐに試せるようになっています。
+API 定義は `utoipa` から生成されるため、ハンドラやスキーマを変更したときは以下を実行して生成物もコミットしてください。
+
+```bash
+makers generate-openapi
+```
+
+CI では同じ生成処理を実行し、`docs/openapi.json` に差分が残っていないことを確認します。
 
 ## 開発環境とミドルウェア
